Now we can log any Object instead of just String

+++ src/main/java/org/newdawn/slick/util/pathfinding/heuristics/ManhattanHeuristic.java
@@ -28,7 +28,7 @@ public class ManhattanHeuristic implements AStarHeuristic {
 	 */
 	public float getCost(TileBasedMap map, Mover mover, int x, int y, int tx,
 			int ty) {
-		return minimumCost * (Math.abs(x-tx) + Math.abs(y-ty));
+		return (float) minimumCost * (Math.abs(x-tx) + Math.abs(y-ty));
 	}
 
 }
